Decided agains result display string conversion
For now it seems to work with '+' and '1s'. Was gona do the string first
then conversion into math, but instruction say otherwise, as to do pairs
of numbers not a full line of like '10+3*7-4' that is done like proper
math, where '*' woud be done first. Also, my way seems waaaay longer,
maybe will play with it later.