Number one and '+' and '=' working
Any number of ones can be input as number one and two. When '+' is
pressed, it's button gets disabled, and input waits for second number.
When that is input, then pressing '=' gives the correct result.
Also, '+' can be pressed multiple times and it doesn't add more of
'+' to mess with anything.