firstNumbNegative() was converting '-' to "+"
the way I wrote it, it was converting the first
numb into a negative as wanted, but then if car-
ried on with subtracting it would convert numbOne
into a positive, as was multipl by '-1'. Added an
'if' statement that checked first if numbOne was
already negative, if so, return.